sudoers_lookup() now returns a bitmap instead of an int.  This makes it possible to express things like "failed to validate because user not listed for this host".  Some thigns that were previously VALIDATE_FOO are now FLAG_FOO.  This may change later on.
Reorganized code in log_auth() and sudo.c to deal with above changes.

Safer versions of push/pushcp with in the do { ... } while (0) style

parse.yacc now saves info on the stack to allow parse.c to determine
if a user was listed, but not for the host he/she tried to run on.

Added --with-mail-if-no-host option