Use a list head struct when storing the semi-circular lists and convert to tail queues in the process.  This will allow us to reverse foreach loops more easily and it makes it clearer which functions expect a list as opposed to a single member.
Add macros for manipulating lists.  Some of these should become functions.

When freeing up a list, just pop off the last item in the queue instead
of going from head to tail.  This is simpler since we don't have to
stash a pointer to the next member, we always just use the last one
in the queue until the queue is empty.

Rename match functions that take a list to have list in the name.
Break cmnd_matches() into cmnd_matches() and cmndlist_matches.