lib: util: make mksig{list,name} tools paths configurable
When cross-compiling, the mksig{list,name} tools get built via the
host-side of the build, and then get installed into a local bin directory
to be run/used to generate the 'siglist.c' and 'signame.c' files for the
target build.

Naturally, the host-side builds and the target-side builds have different
folders, so running ./mksig{list,name} won't work.

To accommodate for this, we add a variable that can be overridden via
env-vars at build time to specify where these 2 tools are placed.
An absolute path can be specified as well, which is great for
cross-building as it can avoid any host-installed mksig{list,name} tools.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>