Fix bad getgroups() pattern
getgroups() is not guaranteed to return the same number of groups
between calls. Because groups are a flat namespace and group numbers
might collide if the system is subscribing to different directories,
some systems deduplicate groups with the same GID. The first call
should be used to retrieve the high watermark needed to allocate
the buffer, the second one should be used to retrieve the group
numbers as well as the actual number of groups we have been given.
Under certain circumstances, the current code causes sudo to read
uninitialized content present in the buffer.