Pass SUDO_EVLOOP_ONCE to del_io_events to prevent sudo hanging at exit on non-BSD systems.
In 994f86b40 the flag SUDO_EVLOOP_ONCE was passed to del_io_events
when called from from pty_close to avoid potential hangs on non-BSD
systems. Later, in 8f694cc22, the flag argument was removed from
del_io_events in favor of a boolean argument determining whether
the function should block (which was the other flag being passed
to it besides SUDO_EVLOOP_ONCE). This change lead to the original
hang fix to be lost and the same - or similar - hang is now
reproducible in sudo.

This commit reintroduces the flag argument to del_io_events with
both the SUDO_EVLOOP_ONCE and SUDO_EVLOOP_NONBLOCK values.

Signed-off-by: Jaroslav Jindrak <dzejrou@gmail.com>