Set ec->term_raw to false even if sudo_term_restore() fails.
Either the fd is not a terminal or we don't have the controlling
terminal.  Either way, we can't know the current status of the
terminal and will need to set to raw mode again (if possible).  Also
make sure to set ec->term_raw to false if sudo_term_raw() fails.