Always disable core dumps when sudo sends itself a fatal signal.
When a command exits due to a fatal signal, sudo will re-send that
signal to itself so the shell does not ignore keyboard-generated
signals.  However, now that sudo disables core dumps by default for
the command, we cannot rely on WCOREDUMP() telling us whether or
not the signal will lead to a core dump.  It is safest to always
disable core dumps before sending the signal to ourself.