Add pam_silent configuration setting
This commit adds a boolean setting 'pam_silent' that controls whether
sudo will pass PAM_SILENT to pam_authenticate(). The default for this
setting is 'true'.

When this setting is set to 'false', pam messages such as faillock
lockout message will be displayed, which is not the case for the
default.