Avoid using ioctl(TIOCNOTTY) in the monitor.
We don't need to revoke the terminal in the monitor, just signal
the foreground process group.  This is more portable and has the
same effect as ioctl(TIOCNOTTY) would on Linux.  Since we now signal
the command from the monitor, there is no reason to forward SIGHUP
from the kernel.  GitHub issue #367.