Fix fd_matches_tty for /dev/console on Linux.
Since Linux 2.1.71, /dev/console is not a mere symlink to the actual
device, but a character device on its own.  When using sudo in a shell
started on a serial console with /dev/console as std{in,out,err},
determining that the controlling terminal (from /proc/self/stat) is
open needs to resolve /dev/console into the actual active console
device, i.e., the last token of /sys/class/tty/console/active.

Signed-off-by: Ignacy Gawędzki <ignacy.gawedzki@green-communications.fr>