sudoedit: fix a potential TOCTOU when writing to the target file
Previously, sudoedit would open the target file read-only, make a
temporary copy, then open the target write-only and overwrite its
contents.  There is a potential TOCTOU if another user has write
permission on a directory in the target file's path and is able to
replace the target file with a hard link to a different file.

To fix this, we now open the target file read-write and keep it
open until it is time to overwrite it with the edited temporary
file.

Reported by Jacco van der Have