runaslist_matches: restore ability of user to only specify a group
The removal of the special case in 8c1559e0 and the subsequent
reworking of runaslist_matches() broke support for "sudo -g group"
and "sudo -u myname -g group" when the user portion of the Runas_Spec
is non-empty and the specified group matches the Runas_Spec.
Thanks to Marc Schoolderman of the sudo-rs project for reporting this.