expand_buf: Allocate "newsize" bytes not "needed" bytes
Sync the implementation of expand_buf() with the version in
logsrv_util.c.  We compute "newsize" by rounding up the number of
needed bytes to the nearest power of 2, but the malloc() call used
the unrounded size.  This is only used for messages coming from
sudo_logsrvd, which are not going to be larger than the initial
buffer, so there is no possibility of an overflow.

Thanks to Joshua Rogers for finding this.