Preserve remaining bytes in the log server read callback functions.
We were not properly handling the case where there were a few bytes
left over after processing the message(s).  Previously, we would
reset the length without moving the remaining bytes to the head of
the buffer.  This could result in a corrupted message.

Thanks to Joshua Rogers for finding this.