verify_peer_identity: Do not pass the resolve flag to validate_hostname()
When acting as a relay (client), the certificate of the server we
are connecting to must match either the DNS name or IP address we
used.  The names in the server's cert should be treated as untrusted
and not be resolved to IP addresses for matching.  This matches the
existing behavior of the sudoers log client.

Thanks to Joshua Rogers for finding this.