Change log ID from a path to UUID + path.
A UUID is now stored for each I/O log that sudo_logsrvd creates.
The log ID in a restart message must now contain a log ID that
matches both the I/O log path and the UUID it contains.
For journaled I/O logs, only the UUID is used.