Pass explicit reason to fmt_alert_message() and fmt_reject_message()
In intercept mode, when a sub-command is rejected by policy we need
to pass in the explicit reason for the failure and not rely on the
reason in the closure (which will be NULL).  Recent changes to
sudo_logsrvd require a reason for the RejectMessage.

Found by the ZeroPath AI Security Engineer <https://zeropath.com>