Embed struct eventlog into struct log_details instead of using a pointer.
In some cases we were allocating a new struct eventlog and others
just attaching a pointer which did not always persist after the
function returns.

In practice this was not a problem.  The eventlog data in log_details
was not being accessed after it went out of scope, but this is
brittle and makes analysis difficult.

Found by the ZeroPath AI Security Engineer <https://zeropath.com>