Add TLS connection error callback function.
For sendlog, this will just break out of the event loop.  For a
logsrvd relay, it will try the next relay server (if any) and return
an error to the client if no connection succeeds.  Note that this
only affects TLS negotiation (which may happen asynchronously) after
the actual socket has been connected.

Found by the ZeroPath AI Security Engineer <https://zeropath.com>