Use a stamp file to mark the multiple outputs of one command
When one command generates multiple outputs, using a naive one-to-two
depency would introduce race condition when doing parallel compilation.
For example, when the rules are written in this pattern:

foo.c foo.h: foo.y
	command foo.y generating foo.c and foo.h
some-target: foo.c
	some command reading foo.c
some-other-target: foo.h
	some command reading foo.h

and the targets are built in parallel, `command foo.y` may be executed
twice because make thinks `command foo.y` can be used to generate foo.c
and can also be used to generate foo.h. What's worse, one of the
commands reading foo.c or foo.h may read an incomplete input during the
command generating the other one is running.

This pattern happens on the rules of def_data.in and gram.y. This
commit fixes this by using a stamp file to witness the execution of the
`command foo.y`, following the advice in the automake manual [1].

[1]: https://www.gnu.org/software/automake/manual/html_node/Multiple-Outputs.html