exec_pty: fix background mode when running the command in a pseudo-tty
The replacement of fd_matches_pgrp() with fd_matches_tty() in
fd3ff3a0bd4ab4f503075e2c4e4e520dbd877f99 resulted in the wrong
branch of the if/else being executed.  Checking for CD_BACKGROUND
first resolves the issue.

Thanks to Bjorn Baron of the sudo-rs project for reporting this.