exec_pty: Fix "sudo -b" when taking input from a pipe or file
In background mode (sudo -b), the sudo process runs in an orphaned
process group so the calls to fd_matches_pgrp() will always return
false.  Move the logic to handle CD_BACKGROUND into the
non-interpose section of the stdin handling.

Also reverse the logic and check for interpose[fd] instead of
!interpose[fd] to improve readability.